fix. 초기예산이월등록 R&U 수정

+++ src/main/java/kr/co/acctmgmt/mapper/BgtICFMapper.java
@@ -17,5 +17,5 @@ public interface BgtICFMapper {
 
 	void deleteBgtICF(BgtICF bgticf);
 	
-	int getSumBgtICFByCoCdAndBgtCd(BgtCD bgtCD);
+	double getSumBgtICFByCoCdAndBgtCd(BgtCD bgtCD);
 }
